==> notebooks/0_fetch_data.ipynb <==
import os
import json
import pandas as pd
from urllib.request import urlretrieve
from urllib.error import HTTPError

root = 'https://raw.githubusercontent.com/TimeWz667/'
pattern_targets = root + 'IndTB_prep/V1.1-warroom/serve/WarRoom/{:s}/targets.csv'
pattern_care = root + 'IndTB_underreport/v0.4/docs/pars/pars_{:s}.json'

def retrieve_data(k, location):
    meta = dict(Name=location['Name'])

    if len(chain) > 1:
        chn = [
            (chain[0], f'State_{chain[0]}'),
            (chain[-2], f'Region_{chain[-2]}'),
            ('India', 'India')
        ]
    else:
        chn = [('India', 'India')]

    for loc, file in chn:
        try:
            path = pattern_targets.format(file)
            urlretrieve(path, f'data/{k}/targets.csv')
        except HTTPError:
            pass
        else:
            meta['Source_Targets'] = {
                'Location': loc,
                'Url': path
            }
            break

    if len(chain) > 1:
        chn = [
            (chain[0], f's_{chain[0]}'),
            (chain[-2], f'r_{chain[-2]}'),
            ('India', 'overall')
        ]
    else:
        chn = [('India', 'overall')]

    for loc, file in chn:
        try:
            path = pattern_care.format(file)
            urlretrieve(path, f'data/{k}/pars_under.json')
        except HTTPError:
            pass
        else:
            meta['Source_Care'] = {
                'Location': loc,
                'Url': path
            }
            break

    with open(f'data/{k}/meta.json', 'w') as f:
        json.dump(meta, f)

def update_pars(js, targets):
    # Care-seeking/Treatment initiation
    txi = targets[targets.Index == "TxI"]
    txi = txi[txi.Year <= 2020]

    txi_pub = txi[txi.Tag == 'Pub']
    txi_pri = txi[txi.Tag == 'Pri']

    txi_pub = (txi_pub.M * txi_pub.N).sum() / txi_pub.N.sum()
    txi_pri = (txi_pri.M * txi_pri.N).sum() / txi_pri.N.sum()

    csi = targets[targets.Index == "PrCSIPub"].M.mean()

    js['exo'] = {
        'p_txi_pub': txi_pub,
        'p_txi_eng': txi_pri,
        'p_txi_pri': txi_pri,
        'p_csi_pub': csi
    }

    # Treatment outcome
    tx_die = targets[targets.Index == 'TxDie']
    tx_succ = targets[targets.Index == 'TxSucc']

    js['txo'] = txo = dict()

    sel = tx_die[tx_die.Tag == 'Pub']
    txi = sel.N.sum()
    txd = (sel.M * sel.N).sum()

    sel = tx_succ[tx_succ.Tag == 'Pub']
    txs = (sel.M * sel.N).sum()
    txl = txi - txs - txd

    txo['pub'] = {'txi': txi, 'txd': txd, 'txs': txs, 'txl': txl}

    sel = tx_die[tx_die.Tag == 'Pri']
    txi = sel.N.sum()
    txd = (sel.M * sel.N).sum()

    sel = tx_succ[tx_succ.Tag == 'Pri']
    txs = (sel.M * sel.N).sum()
    txl = txi - txs - txd

    txo['eng'] = {'txi': txi, 'txd': txd, 'txs': txs, 'txl': txl}
    
    # Prevalence 
    js['prev'] = {k: targets[targets.Index == k].M.mean() for k in ['PrevAsym', 'PrevSym', 'PrevExCS', 'TBLikeUt']}

locations = pd.read_csv('data/locations.csv')
locations = locations.set_index('Location')
locations = {i: dict(row) for i, row in locations.iterrows()}

for k, location in locations.items():
    chain = location['Chain'].split(':')
    os.makedirs(f'data/{k}', exist_ok=True)
    retrieve_data(k, location)

    targets = pd.read_csv(f'data/{k}/targets.csv')
    js = json.load(open(f'data/{k}/pars_under.json', 'r'))
    update_pars(js, targets)
    
    with open(f'data/{k}/pars_under.json', 'w') as f:
        json.dump(js, f)
