==> notebooks/1_attach_pars_txo.ipynb <==
import json
import pandas as pd
from ppa.pars_txo import *

locations = list(pd.read_csv('data/locations.csv').Location)
locations

for loc in locations:
    js = json.load(open(f'data/{loc}/pars_under.json', 'r'))
    targets = pd.read_csv(f'data/{loc}/targets.csv')

    attach_pars_tx(js)
    
    print(loc[:5], '\t ', sum([p['r_txs_pri'] > 0 for p in js['pars']]) / len(js['pars']))
    js['pars'] = [p for p in js['pars'] if p['r_txs_pri'] > 0]
    
    with open(f'data/{loc}/pars_txo.json', 'w') as f:
        json.dump(js, f)
