==> notebooks/2_attach_pars_cas.ipynb <==
import json
import pandas as pd
from ppa.pars_cas import *
from sims_pars import bayes_net_from_script
import numpy as np

locations = list(pd.read_csv('data/locations.csv').Location)
locations

adr = get_adr(pd.read_csv('data/India/targets.csv'))

with open('data/prior.txt', 'r') as f:
    prior = bayes_net_from_script(f.read())

for loc in locations:
    js = json.load(open(f'data/{loc}/pars_txo.json', 'r'))
    attach_pars_cas(js, prior, adr, max_attempt=10)
    
    valid_pars = [p for p in js['pars'] if p['is_valid']]
    n_valid = len(valid_pars)
    n_p = len(js['pars'])
    print(loc[:5], '\t ', f'{n_valid}/{n_p} ({n_valid/n_p:.1%})')
    js['pars'] = valid_pars
    
    with open(f'data/{loc}/pars_cas.json', 'w') as f:
        json.dump(js, f)
