==> notebooks/4_sim_cohort.ipynb <==
import json
import numpy as np
import pandas as pd
import numpy.random as rd
from collections import Counter
from ppa.pars_cas import restructure_pars_cas

locations = list(pd.read_csv('data/locations.csv').Location)

loc = 'India'

pars = json.load(open(f'data/{loc}/pars_cas.json', 'r'))

class DistCompeting:
    def __init__(self, kr):
        self.KeyRates = kr
        
    def __call__(self):
        kr = [(k, rd.exponential(1 / r)) for k, r in self.KeyRates.items()]
        nxt, tte = min(kr, key = lambda x: x[1])
        return nxt, tte
    
class DistSemi:
    def __init__(self, kr):
        self.States = list(kr.keys())
        ps = np.array(list(kr.values()))
        self.Dur = 1 / ps.sum()
        self.Probs = ps * self.Dur

    def __call__(self):
        nxt = rd.choice(self.States, p=self.Probs)
        return nxt, self.Dur

class Agent:
    def __init__(self):
        self.History = list()
        
        self.N_Vis = 0
        self.HasOnset = False
        self.HasDx = False
        self.HasTxI = False
        self.HasTxSucc = False

    def append(self, st, t):
        self.History.append((st, t))


class Simulator:
    def __init__(self, p):
        p = restructure_pars_cas(p)
        self.Pars = p
        self.Samplers = {
            'Asym': DistCompeting({'Sym': p['r_onset'], 'DieUt': p['r_die_asym'], 'SelfCure': p['r_sc']}),
            'Sym': DistCompeting({'CSI': p['r_csi'], 'DieUt': p['r_die_sym'], 'SelfCure': p['r_sc']}),
            'ExCS': DistCompeting({'ReCSI': p['r_recsi'], 'DieUt': p['r_die_sym'], 'SelfCure': p['r_sc']}),
            'TxPub': DistSemi({'Succ': p['r_txs'][0], 'DieTx': p['r_txd'][0], 'LTFU': p['r_txl'][0]}),
            'TxPri': DistSemi({'Succ': p['r_txs'][1], 'DieTx': p['r_txd'][1], 'LTFU': p['r_txl'][1]})
        }
        self.Absorbing = ['DieUt', 'SelfCure', 'DieTx', 'Succ', 'LTFU']
        
    def _sim_an_agent(self):
        p = simulator.Pars
        
        ag = Agent()

        state, ti = 'Asym', 0

        ag.append(state, ti)

        while not (state in self.Absorbing):
            nxt, tte = simulator.Samplers[state]()

            if nxt == 'Sym':
                ag.HasOnset = True
            elif nxt in ['CSI', 'ReCSI']:
                ag.N_Vis += 1
                sector = rd.choice([0, 1, 2], p=p['p_ent'])
                if rd.random() < p['p_dx'][sector]:
                    ag.HasDx = True
                    if rd.random() < p['p_txi'][sector]:
                        ag.HasTxI = True
                        if rd.random() < p['tx_alo'][sector][0]:
                            nxt = 'TxPub'
                        else:
                            nxt = 'TxPri'
                    else:
                        nxt = 'ExCS'
                else:
                    nxt = 'ExCS'
            state = nxt
            ti += tte
            ag.append(state, ti)
        return ag
        
    def simulate(self, n = 100):
        return [self._sim_an_agent() for _ in range(n)]
        
        

p = pars['pars'][4]
simulator = Simulator(p)

ags = simulator.simulate(n = 30)

ts = np.linspace(0, 2, 41)
ts

def count_history(ags, ts):
    ds = []

    for t0 in ts:
        cnt = Counter()
        for ag in ags:
            history = ag.History
            evt, _ = [h for h in history if h[1] <= t0][-1]
            cnt[evt] += 1

        d = {'Time': t0}
        d.update(cnt)
        ds.append(d)

    ds = pd.DataFrame(ds).fillna(0)
    return ds

def 

test_ssm = 11178433
test_naat = 2179976

det_bac = 1173775
det_cdx = 1231040
det_all = det_bac + det_cdx

ppv = 0.50

tp_all = det_all * ppv

sens_ssm, spec_ssm = 0.64, 0.98
sens_naat, spec_naat = 0.8, 0.98
cs = 0.0192
x = cs
test_ssm * (x * sens_ssm + (1 - x) * (1 - spec_ssm)) + test_naat * (x * sens_naat + (1 - x) * (1 - spec_naat))

test_ssm * (1 - spec_ssm) + test_naat * (1 - spec_naat) + x * (test_ssm * (sens_ssm - 1 + spec_ssm) + test_naat * (sens_naat - 1 + spec_naat))

x = (det_bac - test_ssm * (1 - spec_ssm) - test_naat * (1 - spec_naat)) / (test_ssm * (sens_ssm - 1 + spec_ssm) + test_naat * (sens_naat - 1 + spec_naat))
x

tp = test_ssm * (x * sens_ssm) + test_naat * (x * sens_naat)
fp = test_ssm * (1 - x) * (1 - spec_ssm) + test_naat * (1 - x) * (1 - spec_naat)

tp + fp

fn = test_ssm * (x * (1 - sens_ssm)) + test_naat * (x * (1 - sens_naat))
tn = test_ssm * ((1 - x) * spec_ssm) + test_naat * ((1 - x) * spec_naat)
fn, tn

sens_cdx = (tp_all - tp) / fn
fp_all = det_all - tp_all
spec_cdx = 1 - (fp_all - fp) / tn
sens_cdx, spec_cdx

tp1 = test_ssm * (sens_ssm + (1 - sens_ssm) * sens_cdx) + test_naat * (sens_naat + (1 - sens_naat) * sens_cdx)
sens_all = tp1 / (test_ssm + test_naat)

fp1 = test_ssm * ((1 - spec_ssm) + spec_ssm * (1 - spec_cdx)) + test_naat * ((1 - spec_naat) + spec_naat * (1 - spec_cdx))
spec_all = 1 - fp1 / (test_ssm + test_naat)

sens_all, spec_all

x * sens_all / (x * sens_all + (1 - x) * (1 - spec_all))
