==> notebooks/5_calc_pdx.ipynb <==
import numpy as np
from ppa.healthcare.system import get_system
import json
import pandas as pd

N_Test_SSM = 13914911
N_Test_Xpert = 4120552 
N_Det_Pub = 1688427
N_Det_Eng = 733694
N_DetBac = 513050 + 548981
N_DetCDx = 835930 + 1037210
N_Txi_Pub = 1527464
N_Txi_Eng = 554980
N_Pop = 1425775850

PropXpert_Eng = (215594 + 262160) / (3483130 + 2365739)
BacPerXpert_Eng = 68556 / (215594 + 262160)

N_Test_Xpert_Eng = N_Test_Xpert * PropXpert_Eng
N_DetBac_Eng = N_Test_Xpert_Eng * BacPerXpert_Eng

data = {
    'N_DetBac_Pub': N_DetBac - N_DetBac_Eng,
    'N_DetBac_Eng': N_DetBac_Eng,
    'N_Test_SSM_Pub': N_Test_SSM,
    'N_Test_Xpert_Pub': N_Test_Xpert - N_Test_Xpert_Eng,
    'N_Test_Xpert_Eng': N_Test_Xpert_Eng,
    'N_DetCDx_Pub': N_Det_Pub - (N_DetBac - N_DetBac_Eng),
    'N_DetCDx_Eng': N_Det_Eng - N_DetBac_Eng,
    'N_Txi_Pub': N_Txi_Pub,
    'N_Txi_Eng': N_Txi_Eng,
    'DrugTime_Pri': 0.0006037734 * N_Pop
}

data = {k: v / N_Pop for k, v in data.items()}
data

exo = {
    'sens_ssm': 0.64,
    'spec_ssm': 0.98,
    'sens_xpert': 0.85,
    'sens_xpert_ss-': 0.64,
    'spec_xpert': 0.98,
    'p_csi_pub': 0.483,
    'dur_pub': 0.5,
    'p_loss_sputum': 0.15,
    'p_txi_pub': 0.9,
    'p_txi_eng': 0.76,
    'spec_cdx': 0.8
}

with open('data/India/pars_cs_all.json', 'r') as f:
    pars = json.load(f)
    
pars, prev, txo = pars['pars'], pars['prev'], pars['txo']

ds = list()

for p0 in pars:
    p = dict(exo)
    p['sens_cdx'] = p0['sens_cdx']
    p['spec_cdx'] = p0['spec_cdx']

    p['p_ava_ssm_pub'] = p0['p_ava_ssm_pub']
    p['p_ava_xpert_pub'] = p0['p_ava_naat_pub']
    p['p_ava_xpert_eng'] = p0['p_ava_naat_eng']
    p['p_csi_ppm'] = p0['p_csi_ppm']
    p['p_csi_pub'] = p0['p_csi_pub']

    sys = get_system(p)

    d = dict()
    d['P1_Pub'], d['P2_Pub'], d['P3_Pub'], d['P4_Pub'] = sys.Public.Entry
    d['P3_Eng'], d['P4_Eng'] = sys.Engaged.Entry
    for i, alg in enumerate(sys.Public.Algorithms, 1):
        r0 = alg.dx(100, 0)
        bac = r0['N_Det_SSM'] + r0['N_Det_Xpert']
        cdx = r0.TruePos - bac
        pl = r0.FalseNegPreDx
        ng = r0.FalseNeg - pl


        d.update({
            f'P{i}_TP_Bac': bac,
            f'P{i}_TP_CDx': cdx,
            f'P{i}_FN': ng,
            f'P{i}_FN_PDxLTFU': pl
        })

        r1 = alg.dx(0, 100)
        bac = r1['N_Det_SSM'] + r1['N_Det_Xpert']
        cdx = r1.FalsePos - bac
        pl = r1.TrueNegPreDx
        ng = r1.TrueNeg - pl

        d.update({
            f'P{i}_FP_Bac': bac,
            f'P{i}_FP_CDx': cdx,
            f'P{i}_TN': ng,
            f'P{i}_TN_PDxLTFU': pl
        })

    ds.append(d)
    
ds = pd.DataFrame(ds)

ds.describe().T

ds = list()

for p0 in pars:
    p = dict(exo)
    p['sens_cdx'] = p0['sens_cdx']
    p['spec_cdx'] = p0['spec_cdx']

    p['p_ava_ssm_pub'] = p0['p_ava_ssm_pub']
    p['p_ava_xpert_pub'] = p0['p_ava_naat_pub']
    p['p_ava_xpert_eng'] = p0['p_ava_naat_eng']
    p['p_csi_ppm'] = p0['p_csi_ppm']
    p['p_csi_pub'] = p0['p_csi_pub']
    

    sys = get_system(p)

    d = dict()
    d['p_entry_pub'], d['p_entry_eng'], d['p_entry_pri'] = sys.Entry
    d['sens_cdx'] = p0['sens_cdx']
    d['spec_cdx'] = p0['spec_cdx']
    d['p_txi_pub'] = p0['p_txi_pub']
    d['p_txi_eng'] = p0['p_txi_eng']
    d['p_txi_pri'] = p0['p_txi_pri']
    d['p_pri_on_pub'] = p0['p_pri_on_pub']
    
    
    for i, sector, ava in [('Pub', sys.Public, 1 - (1 - p['p_ava_ssm_pub']) * (1 - p['p_ava_xpert_pub'])), ('Eng', sys.Engaged, p0['p_ava_naat_eng']), ('Pri', sys.Private, 1)]:
        r0 = sector.seek_care(100, 0)

        d.update({
            f'{i}_TP': r0.TruePos,
            f'{i}_TP_Bac': (r0['N_Det_SSM'] + r0['N_Det_Xpert']) / ava,
            f'{i}_FN_Bac_Fail': r0['N_SampleFailed'] / ava,
        })

        r1 = sector.seek_care(0, 100)

        d.update({
            f'{i}_FP': r0.FalsePos,
            f'{i}_FP_Bac': (r1['N_Det_SSM'] + r1['N_Det_Xpert']) / ava,
            f'{i}_TN_Bac_Fail': r1['N_SampleFailed'] / ava,
        })
        
    r0 = sys.seek_care(100, 0)
    r0 = (r0['Public'] + r0['Engaged'] + r0['Private'])
    d['All_TP'] = r0.TruePos
    p_txi = np.array([d['p_txi_pub'], d['p_txi_eng'], d['p_txi_pri']])
    d['All_Txi'] = (sys.Entry * np.array([d['Pub_TP'], d['Eng_TP'], d['Pri_TP']]) * p_txi).sum()
    
    ds.append(d)
    
ds = pd.DataFrame(ds)
ds.describe().T

r0 = sys.Public.seek_care(100, 0)
r0.print(True)
