==> notebooks/7_run_cohort.ipynb <==
from sims_pars import bayes_net_from_script, sample
from scipy.integrate import solve_ivp
import numpy as np
import pandas as pd
import tqdm
from ppa.pars import *
from ppa.ebm.cohort import Model

with open('data/prior.txt', 'r') as f:
    prior = bayes_net_from_script(f.read())

cr = CasRepo.load('data/pars_india.json')


m = Model()

mss = list()

for i in tqdm.tqdm(range(1000)):
    exo = sample(prior)
    ps = cr.prepare_pars(exo=exo)
    ms = m.simulate(ps)
    mss.append(ms.assign(Key=i))
    
mss = pd.concat(mss)

mss.to_csv('docs/run_cohort.csv')
