==> notebooks/8_Test_ABM.ipynb <==
import numpy.random as rd
from collections import namedtuple

Result = namedtuple('Result', ('TP', 'FN', 'FP', 'TN'))



class Test:
    def __init__(self, name, sens, spec):
        self.Sensitivity = sens
        self.Specificity = spec
        
    def __call__(self, gp):
        if gp.Active:
            return 
        
    def test(self, gp):
        

Result(TP = 3)

class Hospital:
    def __init__(self.)
